requirements.txt
click>=8.0.0
jinja2>=3.0.0
pyyaml>=6.0
azure-cli>=2.42.0
azure-mgmt-datafactory>=2.0.0
azure-mgmt-storage>=20.0.0
azure-mgmt-keyvault>=10.0.0
azure-mgmt-sql>=3.0.0
azure-mgmt-synapse>=2.0.0
azure-mgmt-resource>=21.0.0
